refactor: inline query planner tests
